🐛 Fix sidebar quick open/close on mobile
