+++ frontend/lib/components/Sidebar/Sidebar.tsx
@@ -1,18 +1,25 @@
+"use client";
 import { motion, MotionConfig } from "framer-motion";
+import { usePathname } from "next/navigation";
+import { useEffect, useState } from "react";
 import { MdChevronRight } from "react-icons/md";
 
+import { useDevice } from "@/lib/hooks/useDevice";
 import { cn } from "@/lib/utils";
+
 type SidebarProps = {
-  open: boolean;
-  setOpen: (open: boolean) => void;
   children: React.ReactNode;
 };
 
-export const Sidebar = ({
-  open,
-  setOpen,
-  children,
-}: SidebarProps): JSX.Element => {
+export const Sidebar = ({ children }: SidebarProps): JSX.Element => {
+  const { isMobile } = useDevice();
+  const pathname = usePathname();
+  const [open, setOpen] = useState(!isMobile);
+
+  useEffect(() => {
+    setOpen(!isMobile);
+  }, [isMobile, pathname, setOpen]);
+
   return (
     <MotionConfig transition={{ mass: 1, damping: 10 }}>
       <motion.div
@@ -29,6 +36,7 @@ export const Sidebar = ({
         className="flex flex-col lg:sticky fixed top-0 left-0 bottom-0 lg:h-[100vh] overflow-visible z-30 border-r border-black/10 dark:border-white/25 bg-white dark:bg-black"
       >
         <motion.div
+          initial={{ display: open ? "flex" : "none" }}
           animate={{
             width: open ? "fit-content" : "0px",
             opacity: open ? 1 : 0.5,
