+++ frontend/lib/hooks/useDevice.ts
@@ -1,12 +1,16 @@
 import { useEffect, useState } from "react";
 
+const MOBILE_MAX_WIDTH = 576;
+
 export const useDevice = (): { isMobile: boolean } => {
-  const [isMobile, setIsMobile] = useState(false);
+  const [isMobile, setIsMobile] = useState(
+    window.innerWidth < MOBILE_MAX_WIDTH
+  );
 
   useEffect(() => {
     const handleResize = () => {
       const screenWidth = window.innerWidth;
-      setIsMobile(screenWidth < 576);
+      setIsMobile(screenWidth < MOBILE_MAX_WIDTH);
     };
 
     // Initial check
